requirements.txt
streamlit==1.27.2
opencv-python==4.10.0.84
numpy==1.26.4
torch==2.5.0
torchvision==0.20.0
Pillow==10.1.0
ultralytics==8.3.23
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
matplotlib==3.9.1
scikit-learn==1.3.2
pyyaml==6.0.1
pickle5==0.0.13
